misc/QR Origen/code/requirements.txt
qrcode~=7.4.2
Pillow~=10.0.1